requirements.txt
azure-functions
-e .